Add atomic_make_stt

+++ test/test_decoders/Audi135/Audi135.cpp
@@ -6,7 +6,7 @@
 static void test_getCrankAngle(void)
 {
   extern decoder_status_t decoderStatus;
-  extern volatile unsigned long toothLastToothTime;
+  extern volatile uint32_t toothLastToothTime;
   extern volatile int toothCurrentCount;
   extern volatile bool revolutionOne;
